Merge pull request #6 from kevinyby/develop

+++ src/com/xinyuan/model/Warehouse/WHScrapOrder.java
@@ -26,7 +26,7 @@ public class WHScrapOrder extends App4 {
 	private String productCode; //产品代码
 	private String productName; //品   名
 	
-	private int amount;//数量
+	private float amount;//数量
 	
 	private Date scrapTime;//报废时间
 	private String scrapReason;//报废理由
@@ -45,10 +45,10 @@ public class WHScrapOrder extends App4 {
 	}
 	
 	
-	public int getAmount() {
+	public float getAmount() {
 		return amount;
 	}
-	public void setAmount(int amount) {
+	public void setAmount(float amount) {
 		this.amount = amount;
 	}
 	public Date getScrapTime() {
